requirements.txt
streamlit
sentence-transformers
farm-haystack
torch
transformers
